+++ .github/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Show credit username during approval process
+
 ...
 
 ## [3.1.0] - 2024-02-18